Got back navigation working
except returning to main activity. NOTE: it is not enabled right now. I
am not going to delete the code that Jayden put in until we talked
about. Right now you cant see the changes I made. If you want to see
that you have to go into Lifestyle/Reminder/Notification activities and
comment out the code I marked in onOptionsItemSelected. The difference
you'll see is now the switch works in the activities, where before it
kinda worked, but not completely.