Array Adapters Completed
Pretty sure the array adapters are done, there is a little more testing
to be done, but I believe it is correct. Failed Reminders/Notifications
don't show up in a Lifestyle/Reminder Array Adapter list now, that was
the way I had to do it, in order to make deletion work properly.